Move chai-extras.js logic to new ChaiLoader class and load via DI.
